fix connections
